@@ -12,41 +12,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.3",
-    "@angular/cdk": "^7.3.0",
-    "@angular/common": "^7.2.3",
-    "@angular/compiler": "^7.2.3",
-    "@angular/core": "^7.2.3",
-    "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "^7.2.3",
-    "@angular/http": "^7.2.3",
-    "@angular/material": "^7.3.0",
-    "@angular/platform-browser": "^7.2.3",
-    "@angular/platform-browser-dynamic": "^7.2.3",
-    "@angular/router": "^7.2.3",
+    "@angular/animations": "^7.2.5",
+    "@angular/cdk": "^7.3.2",
+    "@angular/common": "^7.2.5",
+    "@angular/compiler": "^7.2.5",
+    "@angular/core": "^7.2.5",
+    "@angular/forms": "^7.2.5",
+    "@angular/http": "^7.2.5",
+    "@angular/material": "^7.3.2",
+    "@angular/platform-browser": "^7.2.5",
+    "@angular/platform-browser-dynamic": "^7.2.5",
+    "@angular/router": "^7.2.5",
     "alife-rating-star": "0.0.1",
     "aws-amplify": "^1.1.19",
     "aws-amplify-angular": "^2.1.6",
-    "bulma": "^0.7.2",
-    "core-js": "^2.6.3",
-    "hammerjs": "^2.0.8",
-    "jasmine": "^3.3.1",
+    "bulma": "^0.7.4",
+    "core-js": "^2.6.5",
     "moment": "^2.24.0",
     "ng2-completer": "^2.0.8",
-    "popper.js": "^1.14.7",
     "rxjs": "^6.4.0",
     "rxjs-compat": "^6.4.0",
-    "webpack": "^4.29.0",
+    "webpack": "^4.29.4",
     "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.0",
-    "@angular/cli": "^7.3.0",
-    "@angular/compiler-cli": "^7.2.3",
-    "@angular/language-service": "^7.2.3",
-    "@types/jasmine": "^3.3.8",
+    "@angular-devkit/build-angular": "^0.13.1",
+    "@angular/cli": "^7.3.1",
+    "@angular/compiler-cli": "^7.2.5",
+    "@angular/language-service": "^7.2.5",
+    "@types/jasmine": "^3.3.9",
     "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^10.12.21",
+    "@types/node": "^10.12.26",
     "codelyzer": "^4.3.0",
     "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -58,6 +54,6 @@
     "protractor": "^5.4.2",
     "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "^3.1.6"
+    "typescript": "^3.2.2"
   }
 }
